move test_vocab_download_charngram_vectors (test.test_vocab.TestVocab)
