+++ test/test_build.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python3
 """Tests that requires external resources (Network access to fetch dataset)"""
+from collections import Counter
 
+import numpy as np
 import torch
 import torchtext.data
 
@@ -126,3 +128,36 @@ class TestVocab(TorchtextTestCase):
         token_one_vec = vec.get_vecs_by_tokens(tokens[0], lower_case_backup=True).numpy()
         self.assertEqual(token_one_vec.shape[0], vec.dim)
         torch.testing.assert_allclose(vec[tokens[0].lower()].numpy(), token_one_vec)
+
+
+    def test_vocab_download_charngram_vectors(self):
+        c = Counter({'hello': 4, 'world': 3, 'ᑌᑎIᑕOᗪᕮ_Tᕮ᙭T': 5, 'freq_too_low': 2})
+        # Build a vocab and get vectors twice to test caching, then once more
+        # to test string aliases.
+        for i in range(3):
+            if i == 2:
+                vectors = "charngram.100d"
+            else:
+                vectors = torchtext.vocab.CharNGram()
+            v = torchtext.vocab.Vocab(
+                c, min_freq=3, specials=['<unk>', '<pad>', '<bos>'], vectors=vectors)
+            expected_itos = ['<unk>', '<pad>', '<bos>',
+                             'ᑌᑎIᑕOᗪᕮ_Tᕮ᙭T', 'hello', 'world']
+            expected_stoi = {x: index for index, x in enumerate(expected_itos)}
+            self.assertEqual(v.itos, expected_itos)
+            self.assertEqual(dict(v.stoi), expected_stoi)
+            vectors = v.vectors.numpy()
+
+            # The first 5 entries in each vector.
+            expected_charngram = {
+                'hello': [-0.44782442, -0.08937783, -0.34227219,
+                          -0.16233221, -0.39343098],
+                'world': [-0.29590717, -0.05275926, -0.37334684, 0.27117205, -0.3868292],
+            }
+
+            for word in expected_charngram:
+                torch.testing.assert_allclose(
+                    vectors[v.stoi[word], :5], expected_charngram[word])
+
+            torch.testing.assert_allclose(vectors[v.stoi['<unk>']], np.zeros(100))
+            torch.testing.assert_allclose(vectors[v.stoi['OOV token']], np.zeros(100))
